Added standalone plugins to pom.xml Made Main class for standalone debugging. made example3.yaml Added local Chorme accessibility for selenium scripts.

+++ src/main/java/com/kytheralabs/SeleniumScripter.java
@@ -277,7 +277,7 @@ public class SeleniumScripter {
     private void runWait(Map<String, Object> script) throws Exception {
         int waittimeout = 30;
         if(script.containsKey("timeout")){
-            waittimeout = ((Double) script.get("timeout")).intValue();
+            waittimeout = ((Integer) script.get("timeout")).intValue();
         }
